+++ workspace/plugins/custom-fonts/src/index.ts
@@ -16,23 +16,18 @@
  */
 
 import siyuan from "siyuan";
-
+import { Client } from "@siyuan-community/siyuan-sdk";
 import {
     FLAG_MOBILE,
 } from "@workspace/utils/env/front-end";
 import { Logger } from "@workspace/utils/logger";
-import {
-    removeElementById,
-    moveElementToHeadById,
-} from "@workspace/utils/dom";
-import {
-    updateStyleById,
-} from "@workspace/utils/dom/style";
 import { mergeIgnoreArray } from "@workspace/utils/misc/merge";
 import { classify } from "@workspace/utils/font";
 
 import { type IListItem } from "@workspace/components/siyuan/list/list";
 import List from "@workspace/components/siyuan/list/List.svelte"
+
+import { fontData2CssFontStyle } from "@workspace/utils/font/css";
 
 import Settings from "./components/Settings.svelte";
 
@@ -47,8 +42,9 @@ export default class CustomFontsPlugin extends siyuan.Plugin {
 
     public readonly siyuan = siyuan;
     public readonly logger: InstanceType<typeof Logger>;
+    public readonly client: InstanceType<typeof Client>;
 
-    protected readonly STYLE_ELEMENT_ID: string;
+    protected readonly STYLE_SNIPPET_NAME: string;
     protected readonly SETTINGS_DIALOG_ID: string;
     protected readonly SYSTEM_FONTS_DIALOG_ID: string;
     protected readonly USABLE_FONTS_DIALOG_ID: string;
@@ -59,7 +55,9 @@ export default class CustomFontsPlugin extends siyuan.Plugin {
         super(options);
 
         this.logger = new Logger(this.name);
-        this.STYLE_ELEMENT_ID = `plugin-${this.name}-style`;
+        this.client = new Client();
+
+        this.STYLE_SNIPPET_NAME = `plugin-${this.name}-style`;
         this.SETTINGS_DIALOG_ID = `plugin-${this.name}-settings-dialog`;
         this.SYSTEM_FONTS_DIALOG_ID = `plugin-${this.name}-system-fonts-dialog`;
         this.USABLE_FONTS_DIALOG_ID = `plugin-${this.name}-usable-fonts-dialog`;
@@ -77,11 +75,22 @@ export default class CustomFontsPlugin extends siyuan.Plugin {
     }
 
     onLayoutReady(): void {
-        moveElementToHeadById(this.STYLE_ELEMENT_ID);
     }
 
     onunload(): void {
-        removeElementById(this.STYLE_ELEMENT_ID);
+        this.client.getSnippet({
+            type: "css",
+            enabled: 2,
+        }).then(response => {
+            const snippets = response.data.snippets;
+            const snippet = snippets.find(s => s.name === this.STYLE_SNIPPET_NAME);
+            if (snippet) {
+                snippet.enabled = false;
+                this.client.setSnippet({
+                    snippets,
+                });
+            }
+        });
     }
 
     openSetting(): void {
@@ -103,20 +112,41 @@ export default class CustomFontsPlugin extends siyuan.Plugin {
     }
 
     /* 更新样式 */
-    public updateStyle(): void {
-        updateStyleById(
-            this.STYLE_ELEMENT_ID,
-            [
-                this.config.css.enable ? this.config.css.code : undefined,
-                fontFamilyStyle({
-                    base: this.config.fonts.base.enable ? this.config.fonts.base.list : undefined,
-                    code: this.config.fonts.code.enable ? this.config.fonts.code.list : undefined,
-                    graph: this.config.fonts.graph.enable ? this.config.fonts.graph.list : undefined,
-                    math: this.config.fonts.math.enable ? this.config.fonts.math.list : undefined,
-                    emoji: this.config.fonts.emoji.enable ? this.config.fonts.emoji.list : undefined,
-                }),
-            ].join("\n\n"),
-        );
+    public async updateStyle(): Promise<void> {
+        const response = await this.client.getSnippet({
+            type: "css",
+            enabled: 2,
+        });
+        const snippets = response.data.snippets;
+        const snippet = snippets.find(s => s.name === this.STYLE_SNIPPET_NAME)
+        const content = [
+            this.config.css.enable ? this.config.css.code : undefined,
+            fontFamilyStyle({
+                base: this.config.fonts.base.enable ? this.config.fonts.base.list : undefined,
+                code: this.config.fonts.code.enable ? this.config.fonts.code.list : undefined,
+                graph: this.config.fonts.graph.enable ? this.config.fonts.graph.list : undefined,
+                math: this.config.fonts.math.enable ? this.config.fonts.math.list : undefined,
+                emoji: this.config.fonts.emoji.enable ? this.config.fonts.emoji.list : undefined,
+            }),
+        ].join("\n\n");
+
+        if (snippet) {
+            snippet.enabled = true;
+            snippet.content = content;
+        }
+        else {
+            snippets.push({
+                id: siyuan.Lute.NewNodeID(),
+                type: "css",
+                name: this.STYLE_SNIPPET_NAME,
+                memo: "",
+                content,
+                enabled: true,
+            });
+        }
+        await this.client.setSnippet({
+            snippets,
+        });
     }
 
     /* 重置插件配置 */
@@ -189,11 +219,12 @@ export default class CustomFontsPlugin extends siyuan.Plugin {
                     meta: reaular_font.family,
                     style: `font-family: "${reaular_font.family}"`,
                     fold: true,
+                    // indent: "1em",
                     children: font_list.map(font => ({
                         icon: "#iconFont",
                         text: font.fullName,
                         meta: font.style,
-                        style: `font-family: "${font.family}"`,
+                        style: `font: ${fontData2CssFontStyle(font, 14)};`,
                     })),
                 });
             });